+++ src/components/shared/Form/index.ts
@@ -3,7 +3,6 @@ export { default as SegmentedControl } from './SegmentedControl';
 export { default as SettingsSwitch } from './SettingsSwitch';
 export { default as SettingsSlider } from './SettingsSlider';
 export { default as ActionButton }    from './ActionButton';
-export { default as DirectionPicker } from './DirectionPicker';
 export { default as IconButton }      from './IconButton';
 export { default as Dropdown }        from './Dropdown';
 export { default as FontSettingsPanel } from './FontSettingsPanel';